MMVv2: Make writeToMmv to be abstract for Instance
